+++ optimum/quanto/tensor/qbits/qbits.py
@@ -18,6 +18,7 @@ import torch
 from packaging import version
 from torch.autograd import Function
 
+from ..function import QuantizedLinearFunction
 from ..qtensor import QTensor, qfallback
 from ..qtype import qint2, qint4, qtype, qtypes
 from .group import grouped_shape, ungroup
@@ -259,15 +260,14 @@ class QBitsTensor(QTensor):
         During the execution of the standard torch function, a second-level of dispatch will
         happen, but this time directly on individual torch Tensor operations (mainly ATEN).
         """
-        from ..qtensor_func import get_qtensor_func
-
         kwargs = kwargs or {}
+        if func is torch.nn.functional.linear:
 
-        # Look for a func accepting QTensor inputs
-        qfunc = get_qtensor_func(func)
-        if qfunc is not None:
-            return qfunc(*args, **kwargs)
-        # Defer to dispatcher to look instead for QTensor subclasses operations
+            def qlinear(input, other, bias=None):
+                return QuantizedLinearFunction.apply(input, other, bias)
+
+            return qlinear(*args, **kwargs)
+        # Defer to operations dispatcher
         with torch._C.DisableTorchFunctionSubclass():
             return func(*args, **kwargs)
